% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankoff.R
\name{mpr.sankoff}
\alias{mpr.sankoff}
\title{Ancestral character state reconstruction using maximum parsimony}
\usage{
mpr.sankoff(phy, data, cost, levels, ambig)
}
\arguments{
\item{phy}{An object of class \code{tree}.}

\item{data}{A named vector of character state data. If \code{levels} is
this should be a \code{factor} or \code{integer} vector.}

\item{cost}{A matrix with state-to-state transition costs.}

\item{levels}{A vector containing the different states the character can
take. Optional.}

\item{ambig}{A named list containing ambiguous character state encodings.
The names of the list correspond to ambiguous character states. The values
for each list item refer to states in \code{levels}. Thus, if one of the
names in \code{ambig} appears in \code{data}, that data entry is assumed to
potentially map to any of the states in \code{levels} that appear in the
corresponding \code{ambig} entry. Optional.}
}
\value{
A list with three components:
\describe{
\item{mpr}{A matrix containing optimal state sets for each internal node.
Each column represents a state and is either 0 (the optimal state set does
not include the state) ) or 1 (the optimal state set includes the state).}
\item{mpr_count}{The number of maximum parsimony reconstructions.}
\item{simulate}{A function to randomly sample maximum parsimony 
reconstructions.}
}
}
\description{
Perform a maximum parsimony reconstruction of ancestral character
states using Sankoffs's algorithm
}
